finished following and followers services

+++ ForeScore.Services/FollowingServices.cs
@@ -1,4 +1,5 @@
 ﻿using ForeScore.Data;
+using ForeScore.Models.FollowerModels;
 using ForeScore.Models.FollowingModels;
 using System;
 using System.Collections.Generic;
@@ -17,6 +18,9 @@ namespace ForeScore.Services
             _userId = userId;
         }
 
+        //User inputs an email address for whom they want to follow. If the Follower's email is valid, a following entity will be added
+        //which intakes the user's id and the user they are wanting to follow's email and name. The user's information will then be added as a follow eneity taking in
+        //their email and name and the user they are following's Id
         public bool FollowCreate(FollowingAdd model)
         {
             using (var ctx = new ApplicationDbContext())
@@ -26,7 +30,7 @@ namespace ForeScore.Services
                     .Users
                     .Single(e => e.Email == model.Email);
 
-                if(entity == null)
+                if (entity == null)
                 {
                     return false;
                 }
@@ -40,18 +44,38 @@ namespace ForeScore.Services
                 };
 
                 ctx.Following.Add(following);
+
+                var follower =
+                    ctx
+                    .Users
+                    .Single(e => e.Id == _userId);
+
+                AddFollower(follower, entity.Id);
+
                 return ctx.SaveChanges() == 1;
             }
         }
 
-        public bool FollowDelete(FollowingRemove model)
+        public bool FollowDelete(string email)
         {
             using (var ctx = new ApplicationDbContext())
             {
                 var entity =
                     ctx
                     .Following
-                    .Single(e => e.Email == model.Email);
+                    .Single(e => e.Email == email && _userId == e.Id);
+
+                if (entity == null)
+                {
+                    return false;
+                }
+
+                var followed =
+                ctx
+                .Users
+                .Single(e => e.Id == _userId);
+
+                RemoveFollower(followed, entity.Id);
 
 
                 ctx.Following.Remove(entity);
@@ -78,5 +102,41 @@ namespace ForeScore.Services
             }
         }
 
+
+        private bool AddFollower(ApplicationUser model, string id)
+        {
+            using (var ctx = new ApplicationDbContext())
+            {
+                var entity = new FollowedBy();
+
+                entity.Email = model.Email;
+                entity.FullName = model.FullName;
+
+                entity.Id = id;
+
+                ctx.FollowedBy.Add(entity);
+
+                return ctx.SaveChanges() == 1;
+            }
+        }
+
+        private bool RemoveFollower(ApplicationUser model, string id)
+        {
+            using (var ctx = new ApplicationDbContext())
+            {
+                var entity =
+                      ctx
+                      .FollowedBy
+                      .Single(e => id == e.Id && model.Email == e.Email);
+
+                if (entity == null)
+                {
+                    return false;
+                }
+
+                ctx.FollowedBy.Remove(entity);
+                return ctx.SaveChanges() == 1;
+            }
+        }
     }
 }
